Subdir in example

+++ desword/main.py
@@ -1,7 +1,8 @@
 import glob
 import os
 import sys
-from parsing.md import CustomMarkdownParser
+from md import CustomMarkdownParser
+from pathlib import Path
 
 input = sys.argv[1]
 output = sys.argv[2]
@@ -9,9 +10,20 @@ output = sys.argv[2]
 if not output.strip():
     raise Exception('Output blank')
 
+if not output[-1] == os.path.sep:
+    output = f"{output}{os.path.sep}"
+
 files = glob.glob(f"{output}*")
+print(files)
+directories = []
 for f in files:
-    os.remove(f)
+    if not os.path.isfile(f):
+        directories.append(f)
+    else:
+        os.remove(f)
+
+for d in directories:
+    os.rmdir(d)
 
 m = CustomMarkdownParser(output)
 
@@ -28,7 +40,12 @@ class Page:
             self.html += f'\nFrom <a href="{k}">{k}</a> link: {v}'
 
 
-for root, _, files in os.walk(input):
+for root, dirs, files in os.walk(input):
+    input_relative_path = os.path.relpath(
+        root, input)
+    output_relative_path = os.path.join(
+        output, input_relative_path)
+    Path(output_relative_path).mkdir(parents=True, exist_ok=True)
     for file in files:
         file_parts = os.path.splitext(file)
         # Ignore non-markdown files
@@ -37,7 +54,8 @@ for root, _, files in os.walk(input):
             continue
 
         input_location = os.path.join(root, file)
-        output_location = os.path.join(output, file_parts[0] + ".html")
+        output_location = os.path.join(
+            output_relative_path, file_parts[0] + ".html")
         with open(input_location) as f:
             lines = f.read()
         page_graph[output_location] = Page(m, lines)
